MdePkg/BasePeCoffLib2: Fix ARM Thumb apply relocation routine
Thumb instructions are either 16-bit or 32-bit, and are aligned on
a two-byte boundary

+++ MdePkg/Library/BasePeCoffLib2/PeCoffRelocate.c
@@ -379,9 +379,9 @@ InternalApplyRelocation (
       }
       //
       // Verify the Base Relocation target is sufficiently aligned.
-      // The ARM Thumb instruction pair must start on a 32-bit boundary.
+      // The ARM Thumb instruction pair must start on a 16-bit boundary.
       //
-      if (!IS_ALIGNED (RelocTargetRva, ALIGNOF (UINT32))) {
+      if (!IS_ALIGNED (RelocTargetRva, ALIGNOF (UINT16))) {
         DEBUG_RAISE ();
         return RETURN_VOLUME_CORRUPTED;
       }
